FINISHED! BOUNCING BULLET

+++ Makefile
@@ -1,3 +1,8 @@
+.SUFFIXES:
+SUFFIXES =
+.SUFFIXES: .c .cpp .h .hpp .rc .res .inl .o .d .asm
+
+
 #==============================================================================
 MAKEFLAGS += --no-print-directory
 #==============================================================================
@@ -10,6 +15,12 @@ ifeq ($(BUILD),Tests)
 	_BUILDL := release
 endif
 
+# The sub-folder containing the target source files
+SRC_TARGET?=
+ifneq ($(SRC_TARGET),)
+	_SRC_TARGET := /$(SRC_TARGET)
+endif
+
 # Maximum parallel jobs during build process
 MAX_PARALLEL_JOBS?=8
 
@@ -19,11 +30,22 @@ DUMP_ASSEMBLY?=false
 # Clean output?
 CLEAN_OUTPUT?=true
 
+# If dll, build as a static library?
+BUILD_STATIC?=false
+
 # Platform specific environment variables
 -include env/.all.mk
 -include env/.$(_BUILDL).mk
 -include env/$(PLATFORM).all.mk
 -include env/$(PLATFORM).$(_BUILDL).mk
+
+# Target specific variables
+ifneq ($(SRC_TARGET),)
+-include env/$(SRC_TARGET)/.all.mk
+-include env/$(SRC_TARGET)/.$(_BUILDL).mk
+-include env/$(SRC_TARGET)/$(PLATFORM).all.mk
+-include env/$(SRC_TARGET)/$(PLATFORM).$(_BUILDL).mk
+endif
 
 #==============================================================================
 # File/Folder dependencies for the production build recipe (makeproduction)
@@ -59,7 +81,7 @@ NAME?=game.exe
 
 #==============================================================================
 # The source file directory
-SRC_DIR := src
+SRC_DIR := src$(_SRC_TARGET)
 LIB_DIR := lib
 
 # Project .cpp or .rc files (relative to $(SRC_DIR) directory)
@@ -68,7 +90,6 @@ SOURCE_FILES := $(patsubst $(SRC_DIR)/%,%,$(shell find $(SRC_DIR) -name '*.cpp' 
 PROJECT_DIRS := $(patsubst $(SRC_DIR)/%,%,$(shell find $(SRC_DIR) -mindepth 1 -maxdepth 99 -type d))
 
 # Add prefixes to the above variables
-_LIB_DIRS := $(LIB_DIR:%=-L%/) $(LIB_DIRS:%=-L%)
 _INCLUDE_DIRS := $(patsubst %,-I%,$(SRC_DIR)/ $(LIB_DIR)/ $(INCLUDE_DIRS))
 
 _BUILD_MACROS := $(BUILD_MACROS:%=-D%)
@@ -127,22 +148,33 @@ ifeq ($(BUILD),Tests)
 	BLD_DIR := bin/Release
 endif
 BLD_DIR := $(BLD_DIR:%/=%)
+_BASENAME := $(basename $(NAME))
+ifeq ($(BUILD_STATIC),true)
+ifeq ($(suffix $(NAME)),$(filter $(suffix $(NAME)),.dll .dylib .so))
+ifneq ($(suffix $(NAME)),)
+	NAME := $(_BASENAME)-s.a
+endif
+endif
+endif
 TARGET := $(BLD_DIR)/$(NAME)
-_NAMENOEXT := $(NAME:.exe=)
-_NAMENOEXT := $(_NAMENOEXT:.dll=)
 
-_SOURCES_IF_RC := $(if $(filter windows,$(PLATFORM)),$(SOURCE_FILES:.rc=.res),$(SOURCE_FILES:%.rc=))
+ifneq ($(SRC_TARGET),)
+	LIB_DIRS := $(LIB_DIRS) $(BLD_DIR)
+endif
+_LIB_DIRS := $(LIB_DIR:%=-L%/) $(LIB_DIRS:%=-L%)
 
-OBJ_DIR := $(BLD_DIR)/obj
+_SOURCES_IF_RC := $(if $(filter windows,$(PLATFORM)),$(SOURCE_FILES),$(SOURCE_FILES:%.rc=))
+
+OBJ_DIR := $(BLD_DIR)/obj$(_SRC_TARGET)
 _OBJS := $(_SOURCES_IF_RC:.c=.c.o)
 _OBJS := $(_OBJS:.cpp=.cpp.o)
 _OBJS := $(_OBJS:.cc=.cc.o)
+_OBJS := $(_OBJS:.rc=.res)
 OBJS := $(_OBJS:%=$(OBJ_DIR)/%)
 OBJ_SUBDIRS := $(PROJECT_DIRS:%=$(OBJ_DIR)/%)
 
-DEP_DIR := $(BLD_DIR)/dep
-_DEPS := $(_SOURCES_IF_RC)
-_DEPS := $(_DEPS:%=%.d)
+DEP_DIR := $(BLD_DIR)/dep$(_SRC_TARGET)
+_DEPS := $(_SOURCES_IF_RC:%=%.d)
 DEPS := $(_DEPS:%=$(DEP_DIR)/%) $(DEP_DIR)/$(PRECOMPILED_HEADER).d
 DEP_SUBDIRS := $(PROJECT_DIRS:%=$(DEP_DIR)/%)
 
@@ -151,6 +183,8 @@ _PCH_HFILE := $(_PCH_HFILE:$(SRC_DIR)/%=%)
 _PCH_EXT := $(_PCH_HFILE:$(PRECOMPILED_HEADER).%=%)
 _PCH_COMPILER_EXT := $(if $(filter osx,$(PLATFORM)),p,g)ch
 
+_SYMBOLS := $(if $(filter osx,$(PLATFORM)),,$(if $(filter Release,$(BUILD)),-s,))
+
 
 _PCH := $(_PCH_HFILE:%=$(OBJ_DIR)/%)
 ifneq ($(_PCH),)
@@ -158,7 +192,7 @@ ifneq ($(_PCH),)
 endif
 
 ifeq ($(DUMP_ASSEMBLY),true)
-	ASM_DIR := $(BLD_DIR)/asm
+	ASM_DIR := $(BLD_DIR)/asm$(_SRC_TARGET)
 	_ASMS := $(_OBJS:%.res=)
 	_ASMS := $(_ASMS:.o=.o.asm)
 	ASMS := $(_ASMS:%=$(ASM_DIR)/%)
@@ -168,6 +202,8 @@ endif
 _DIRECTORIES := $(sort bin $(BLD_DIR) $(OBJ_DIR) $(OBJ_SUBDIRS) $(DEP_DIR) $(DEP_SUBDIRS) $(ASM_DIR) $(ASM_SUBDIRS))
 
 _CLEAN := $(filter true,$(CLEAN_OUTPUT))
+
+_TARGET_DEPS := $(_PCH_GCH) $(OBJS) $(ASMS) $(TEST_DIR)
 
 # Quiet flag
 _Q := $(if $(_CLEAN),@)
@@ -188,7 +224,7 @@ endif
 OBJ_COMPILE = $(CC) $(CFLAGS_DEPS) $(_BUILD_MACROS) $(_INCLUDE_DIRS) $(_INCLUDE_PCH) $(CFLAGS) -o $@ -c $<
 OBJ_COMPILE_T = $(CC) $(CFLAGS_DEPS_T) $(_BUILD_MACROS) $(_INCLUDE_DIRS) $(_INCLUDE_PCH) $(CFLAGS) -o $@ -c $<
 
-RC_COMPILE = -$(RC) -J rc -O coff -i $< -o $@
+RC_COMPILE = -$(RC) -J rc -O coff --preprocessor-arg=-MT --preprocessor-arg=$@ --preprocessor-arg=-MMD --preprocessor-arg=-MP --preprocessor-arg=-MF --preprocessor-arg=$(DEP_DIR)/$*.rc.Td $(_BUILD_MACROS) $(_INCLUDE_DIRS) -i $< -o $@
 ifeq ($(PLATFORM),osx)
 	ASM_COMPILE = otool -tvV $< | c++filt > $@
 else
@@ -196,116 +232,186 @@ else
 endif
 POST_COMPILE = mv -f $(DEP_DIR)/$*.Td $(DEP_DIR)/$*.d && touch $@
 POST_COMPILE_T = mv -f $(DEP_DIR)/.$(TEST_DIR)/$*.Td $(DEP_DIR)/.$(TEST_DIR)/$*.d && touch $@
+POST_COMPILE_RC = mv -f $(DEP_DIR)/$*.rc.Td $(DEP_DIR)/$*.rc.d && touch $@
+
+#==============================================================================
+# Unicode
+UNI_COPY := echo -n "➦"
+UNI_LINK := echo -n "⇛"
+ifeq ($(PLATFORM),osx)
+	UNI_COPY := printf "➦"
+	UNI_LINK := printf "⇛"
+endif
+ifeq ($(PLATFORM),windows)
+	UNI_COPY := printf '\xE2\x9E\xA6'
+	UNI_LINK := printf '\xE2\x87\x9B'
+endif
+
+# Misc
+ORIGIN_FLAG := '-Wl,-R$$ORIGIN'
+ifeq ($(PLATFORM),osx)
+	ORIGIN_FLAG :=
+endif
 
 #==============================================================================
 # Build Scripts
-.DELETE_ON_ERROR: all
 all:
 	@$(MAKE) makepch
 	@$(MAKE) -j$(MAX_PARALLEL_JOBS) makebuild
+.DELETE_ON_ERROR: all
 
-.PHONY: rebuild
 rebuild: clean all
+.PHONY: rebuild
 
-.PHONY: buildprod
 buildprod: all makeproduction
+.PHONY: buildprod
 
 #==============================================================================
 # Functions
-define color_reset
-	@tput setaf 4
-endef
+color_reset := @tput setaf 4
 
-define comple_with
+define compile_with
 	$(color_reset)
-	$(if $(_CLEAN),@echo $($(2):$(OBJ_DIR)/%=%))
+	$(if $(_CLEAN),@echo '   $($(2):$(OBJ_DIR)/%=%)')
 	$(_Q)$(3) && $(4)
 endef
 
+define linking_with
+	$(color_reset)
+	$(if $(_CLEAN),@echo; $(UNI_LINK); echo '  Linking: $(1)')
+endef
+
+define build_deps
+	$(foreach dep,$(BUILD_DEPENDENCIES),$(call copy_to,$(dep),$(BLD_DIR)))
+endef
+
 MKDIR := $(_Q)mkdir -p
+
+makepch: $(_PCH_GCH)
+	@echo > /dev/null
+.PHONY: makepch
+
+makebuild: $(TARGET)
+	$(color_reset)
+ifeq ($(SRC_TARGET),)
+	@echo '   Target is up to date.'
+else
+	@echo '   $(NAME): Target is up to date.'
+endif
+.PHONY: makebuild
 
 #==============================================================================
 # Build Recipes
 $(OBJ_DIR)/%.o: $(SRC_DIR)/%
 $(OBJ_DIR)/%.o: $(SRC_DIR)/% $(_PCH_GCH) $(DEP_DIR)/%.d | $(_DIRECTORIES)
-	$(call comple_with,@,<,$(OBJ_COMPILE),$(POST_COMPILE))
+	$(call compile_with,@,<,$(OBJ_COMPILE),$(POST_COMPILE))
 
 $(OBJ_DIR)/.$(TEST_DIR)/%.o: $(TEST_DIR)/%
 $(OBJ_DIR)/.$(TEST_DIR)/%.o: $(TEST_DIR)/% $(_PCH_GCH) $(DEP_DIR)/.$(TEST_DIR)/%.d | $(_DIRECTORIES)
-	$(call comple_with,@,<,$(OBJ_COMPILE_T),$(POST_COMPILE_T))
+	$(call compile_with,@,<,$(OBJ_COMPILE_T),$(POST_COMPILE_T))
 
 $(OBJ_DIR)/%.$(_PCH_EXT).$(_PCH_COMPILER_EXT) : $(SRC_DIR)/%.$(_PCH_EXT)
 $(OBJ_DIR)/%.$(_PCH_EXT).$(_PCH_COMPILER_EXT) : $(SRC_DIR)/%.$(_PCH_EXT) $(DEP_DIR)/%.d | $(_DIRECTORIES)
-	$(call comple_with,@,<,$(PCH_COMPILE),$(POST_COMPILE))
+	$(call compile_with,@,<,$(PCH_COMPILE),$(POST_COMPILE))
 
 $(OBJ_DIR)/%.res: $(SRC_DIR)/%.rc
-$(OBJ_DIR)/%.res: $(SRC_DIR)/%.rc $(DEP_DIR)/%.d | $(_DIRECTORIES)
-	$(color_reset)
-	$(if $(_CLEAN),@echo $(<:$(OBJ_DIR)/%=%))
-	$(_Q)$(RC_COMPILE)
+$(OBJ_DIR)/%.res: $(SRC_DIR)/%.rc $(DEP_DIR)/%.rc.d | $(_DIRECTORIES)
+	$(call compile_with,@,<,$(RC_COMPILE),$(POST_COMPILE_RC))
 
 $(ASM_DIR)/%.o.asm: $(OBJ_DIR)/%.o
-	$(if $(_CLEAN),,$(color_reset))
+	@tput setaf 6
+	$(if $(_CLEAN),@echo "   $@")
 	$(_Q)$(ASM_COMPILE)
 
-$(TARGET): $(_PCH_GCH) $(OBJS) $(ASMS) $(BLD_DIR) $(TEST_DIR)
-	$(color_reset)
-	$(if $(_CLEAN),@echo; echo 'Linking: $(TARGET)')
-ifeq ($(suffix $(TARGET)),.dll)
-	-$(_Q)rm -rf $(BLD_DIR)/lib$(_NAMENOEXT).def $(BLD_DIR)/lib$(_NAMENOEXT).a
-	$(_Q)$(CC) -shared -Wl,--output-def="$(BLD_DIR)/lib$(_NAMENOEXT).def" -Wl,--out-implib="$(BLD_DIR)/lib$(_NAMENOEXT).a" -Wl,--dll $(_LIB_DIRS) $(OBJS) -o $@ -s $(_LINK_LIBRARIES) $(BUILD_FLAGS)
-else
-	$(_Q)$(CC) $(_LIB_DIRS) -o $@ $(OBJS) $(_LINK_LIBRARIES) $(BUILD_FLAGS)
-endif
-	$(foreach dep,$(BUILD_DEPENDENCIES),$(shell cp -r $(dep) $(BLD_DIR)))
+$(BLD_DIR)/lib%-s.a: $(_TARGET_DEPS)
+	$(call linking_with,$@)
+	-$(_Q)rm -rf $@
+	$(_Q)ar -c -r -s $@ $(OBJS)
+	@echo
 
+$(BLD_DIR)/$(_BASENAME).dll: $(_TARGET_DEPS)
+	$(call linking_with,$@)
+	-$(_Q)rm -rf $(BLD_DIR)/$(_BASENAME).def $(BLD_DIR)/$(_BASENAME).a
+	$(_Q)$(CC) -shared -Wl,--output-def="$(BLD_DIR)/$(_BASENAME).def" -Wl,--out-implib="$(BLD_DIR)/$(_BASENAME).a" -Wl,--dll $(_LIB_DIRS) $(OBJS) -o $@ $(_SYMBOLS) $(_LINK_LIBRARIES) $(BUILD_FLAGS)
+	@echo
 
-.PHONY: makepch
-makepch: $(_PCH_GCH)
-	@echo > /dev/null
+$(BLD_DIR)/$(_BASENAME).so: $(_TARGET_DEPS)
+	$(call linking_with,$@)
+	$(_Q)$(CC) -shared $(_LIB_DIRS) $(OBJS) -o $@ $(_SYMBOLS) $(_LINK_LIBRARIES) $(BUILD_FLAGS)
+	@echo
 
-.PHONY: makebuild
-makebuild: $(TARGET)
-	$(color_reset)
-	@echo '$(BUILD) build target is up to date.'
+$(BLD_DIR)/$(_BASENAME).dylib: $(_TARGET_DEPS)
+	$(call linking_with,$(BLD_DIR)/$(_BASENAME).dylib)
+	$(_Q)$(CC) -dynamiclib -undefined suppress -flat_namespace $(_LIB_DIRS) $(OBJS) -o $@ $(_SYMBOLS) $(_LINK_LIBRARIES) $(BUILD_FLAGS)
+	@echo
+
+$(BLD_DIR)/$(_BASENAME).exe: $(_TARGET_DEPS)
+	$(call linking_with,$@)
+	$(_Q)$(CC) $(_LIB_DIRS) $(_SYMBOLS) -o $@ $(ORIGIN_FLAG) $(OBJS) $(_LINK_LIBRARIES) $(BUILD_FLAGS)
+	@echo
+	$(call build_deps)
+
+$(BLD_DIR)/$(_BASENAME): $(_TARGET_DEPS)
+	$(call linking_with,$@)
+	$(_Q)$(CC) $(_LIB_DIRS) $(_SYMBOLS) -o $@ $(ORIGIN_FLAG) $(OBJS) $(_LINK_LIBRARIES) $(BUILD_FLAGS)
+	@echo
+	$(call build_deps)
 
 $(_DIRECTORIES):
 	$(if $(_CLEAN),,$(color_reset))
 	$(MKDIR) $@
+	$(if $(_CLEAN),,@echo)
 
-.PHONY: clean
 clean:
 	$(color_reset)
-	$(if $(_CLEAN),@echo 'Cleaning old build files & folders...'; echo)
+	$(if $(_CLEAN),@echo '   Cleaning old build files & folders...'; echo)
 	$(_Q)$(RM) $(TARGET) $(DEPS) $(OBJS)
+.PHONY: clean
 
 #==============================================================================
 # Production recipes
 
-.PHONY: rmprod
 rmprod:
 	$(color_reset)
+	@echo
 	-$(_Q)rm -rf $(if $(filter osx,$(PLATFORM)),$(PRODUCTION_FOLDER_MACOS),$(PRODUCTION_FOLDER))
 ifeq ($(PLATFORM),linux)
 	-$(_Q)rm -rf ~/.local/share/applications/$(NAME).desktop
 endif
+.PHONY: rmprod
 
-.PHONY: mkdirprod
 mkdirprod:
 	$(color_reset)
 	$(MKDIR) $(PRODUCTION_FOLDER)
+.PHONY: mkdirprod
 
-.PHONY: releasetoprod
+define do_copy_to_clean
+	@$(UNI_COPY)
+	@echo  "  Copying \"$(1)\" to \"$(CURDIR)/$(2)\""
+	$(shell cp -r $(1) $(2))
+endef
+
+define do_copy_to
+	@echo  "cp -r $(1) $(2)"
+	$(shell cp -r $(1) $(2))
+endef
+
+define copy_to
+	$(if $(wildcard $(2)/$(notdir $(1))),,$(if $(_CLEAN),$(call do_copy_to_clean,$(1),$(2)),$(call do_copy_to,$(1),$(2))))
+endef
+
 releasetoprod: $(TARGET)
 	$(color_reset)
 ifeq ($(PLATFORM),osx)
-	@echo 'Creating the MacOS application bundle...'
+	@echo '   Creating the MacOS application bundle...'
+	@echo
 	$(MKDIR) $(PRODUCTION_FOLDER)/Resources $(PRODUCTION_FOLDER)/Frameworks $(PRODUCTION_FOLDER)/MacOS
 ifeq ($(shell brew ls --versions makeicns),)
 	brew install makeicns
 	$(color_reset)
 endif
 	$(_Q)makeicns -in env/osx/$(PRODUCTION_MACOS_ICON).png -out $(PRODUCTION_FOLDER)/Resources/$(PRODUCTION_MACOS_ICON).icns
+	@echo
 	$(_Q)plutil -convert binary1 env/osx/Info.plist.json -o $(PRODUCTION_FOLDER)/Info.plist
 	$(_Q)plutil -replace CFBundleExecutable -string $(NAME) $(PRODUCTION_FOLDER)/Info.plist
 	$(_Q)plutil -replace CFBundleName -string $(PRODUCTION_MACOS_BUNDLE_NAME) $(PRODUCTION_FOLDER)/Info.plist
@@ -328,24 +434,31 @@ else ifeq ($(PLATFORM),linux)
 	$(_Q)cp $(PRODUCTION_FOLDER)/$(NAME).desktop ~/.local/share/applications
 else
 	$(_Q)cp $(TARGET) $(PRODUCTION_FOLDER)
+	$(if $(_CLEAN),,@echo)
 endif
+.PHONY: releasetoprod
 
-.PHONY: makeproduction
 makeproduction: rmprod mkdirprod releasetoprod
 	$(color_reset)
-	@echo -n 'Adding dynamic libraries & project dependencies...'
-	$(foreach dep,$(PRODUCTION_DEPENDENCIES),$(shell cp -r $(dep) $(PRODUCTION_FOLDER_RESOURCES)))
+ifneq ($(PRODUCTION_DEPENDENCIES),)
+	@echo '   Adding dynamic libraries & project dependencies...'
+	@echo
+	$(foreach dep,$(PRODUCTION_DEPENDENCIES),$(call copy_to,$(dep),$(PRODUCTION_FOLDER_RESOURCES)))
 	$(foreach excl,$(PRODUCTION_EXCLUDE),$(shell find $(PRODUCTION_FOLDER_RESOURCES) -name '$(excl)' -delete))
-	@echo ' Done'
+endif
 ifeq ($(PLATFORM),osx)
-	$(foreach dylib,$(PRODUCTION_MACOS_DYLIBS),$(shell cp -r $(dylib) $(PRODUCTION_FOLDER)/MacOS ))
+	$(foreach dylib,$(PRODUCTION_MACOS_DYLIBS),$(call copy_to,$(dylib),$(PRODUCTION_FOLDER)/MacOS))
 	$(_Q)install_name_tool -add_rpath @executable_path/../Frameworks $(PRODUCTION_FOLDER)/MacOS/$(NAME)
 	$(_Q)install_name_tool -add_rpath @loader_path/.. $(PRODUCTION_FOLDER)/MacOS/$(NAME)
+	$(foreach dylib,$(PRODUCTION_MACOS_DYLIBS),$(shell install_name_tool -change @rpath/$(notdir $(dylib)) @rpath/MacOS/$(notdir $(dylib)) $(PRODUCTION_FOLDER)/MacOS/$(NAME)))
 	$(foreach dylib,$(PRODUCTION_MACOS_DYLIBS),$(shell install_name_tool -change $(notdir $(dylib)) @rpath/MacOS/$(notdir $(dylib)) $(PRODUCTION_FOLDER)/MacOS/$(NAME)))
-	$(foreach framework,$(PRODUCTION_MACOS_FRAMEWORKS),$(shell cp -r $(framework) $(PRODUCTION_FOLDER)/Frameworks))
+	$(foreach dylib,$(PRODUCTION_MACOS_DYLIBS),$(shell install_name_tool -change $(dylib) @rpath/MacOS/$(notdir $(dylib)) $(PRODUCTION_FOLDER)/MacOS/$(NAME)))
+	$(foreach framework,$(PRODUCTION_MACOS_FRAMEWORKS),$(call copy_to,$(framework),$(PRODUCTION_FOLDER)/Frameworks))
 ifeq ($(PRODUCTION_MACOS_MAKE_DMG),true)
 	$(shell hdiutil detach /Volumes/$(PRODUCTION_MACOS_BUNDLE_NAME)/ &> /dev/null)
-	@echo 'Creating the dmg image for the application...'
+	@echo
+	@echo '   Creating the dmg image for the application...'
+	@echo
 	$(_Q)hdiutil create -megabytes 54 -fs HFS+ -volname $(PRODUCTION_MACOS_BUNDLE_NAME) $(PRODUCTION_FOLDER_MACOS)/.tmp.dmg > /dev/null
 	$(_Q)hdiutil attach $(PRODUCTION_FOLDER_MACOS)/.tmp.dmg > /dev/null
 	$(_Q)cp -r $(PRODUCTION_FOLDER_MACOS)/$(PRODUCTION_MACOS_BUNDLE_NAME).app /Volumes/$(PRODUCTION_MACOS_BUNDLE_NAME)/
@@ -357,9 +470,11 @@ ifeq ($(PRODUCTION_MACOS_MAKE_DMG),true)
 	$(_Q)hdiutil detach /Volumes/$(PRODUCTION_MACOS_BUNDLE_NAME)/ > /dev/null
 	$(_Q)hdiutil convert $(PRODUCTION_FOLDER_MACOS)/.tmp.dmg -format UDZO -o $(PRODUCTION_FOLDER_MACOS)/$(PRODUCTION_MACOS_BUNDLE_NAME).dmg > /dev/null
 	$(_Q)rm -f $(PRODUCTION_FOLDER_MACOS)/.tmp.dmg
-	@echo 'Created $(PRODUCTION_FOLDER_MACOS)/$(PRODUCTION_MACOS_BUNDLE_NAME).dmg'
+	@echo
+	@echo '   Created $(PRODUCTION_FOLDER_MACOS)/$(PRODUCTION_MACOS_BUNDLE_NAME).dmg'
 endif
 endif
+.PHONY: makeproduction
 
 #==============================================================================
 # Dependency recipes
